requirements.txt
python-telegram-bot
beautifulsoup4
requests
pandas
lxml
xlrd
CacheControl
tinydb